+++ pennyway-domain/src/main/java/kr/co/pennyway/domain/common/redis/message/type/MessageCategoryType.java
@@ -1,4 +1,4 @@
-package kr.co.pennyway.domain.common.redis.message.domain;
+package kr.co.pennyway.domain.common.redis.message.type;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import kr.co.pennyway.domain.common.converter.LegacyCommonType;
@@ -8,24 +8,22 @@ import java.util.Map;
 import java.util.stream.Stream;
 
 @RequiredArgsConstructor
-public enum MessageContentType implements LegacyCommonType {
-    TEXT("0", "TEXT"),
-    IMAGE("1", "IMAGE"),
-    VIDEO("2", "VIDEO"),
-    FILE("3", "FILE");
+public enum MessageCategoryType implements LegacyCommonType {
+    NORMAL("0", "NORMAL"),
+    SYSTEM("1", "SYSTEM");
 
-    private static final Map<String, MessageContentType> stringToEnum = Stream.of(values()).collect(java.util.stream.Collectors.toMap(Object::toString, e -> e));
+    private static final Map<String, MessageCategoryType> stringToEnum = Stream.of(values()).collect(java.util.stream.Collectors.toMap(Object::toString, e -> e));
     private final String code;
     private final String type;
 
     @JsonCreator
-    public static MessageContentType fromString(String type) {
+    public static MessageCategoryType fromString(String type) {
         return stringToEnum.get(type.toUpperCase());
     }
 
     @Override
     public String getCode() {
-        return code;
+        return null;
     }
 
     @Override
